sharness.sh: export SHELL_PATH
In Git, "test-lib.sh" sources the GIT-BUILD-OPTIONS file
where SHELL_PATH is defined, and then exports SHELL_PATH,
so this environment variable is available to all the
test scripts.

In sharness there is no GIT-BUILD-OPTIONS, so "sharness.sh"
doesn't have SHELL_PATH defined, and doesn't export it.

This is bad because SHELL_PATH is used in "sharness.t".

Now "sharness.t" still works because the following works:

$ cat >test.sh <<EOF
echo "it works!"
EOF
$ chmod +x test.sh
$ ./test.sh
it works!

But it's still better in "sharness.sh" to define SHELL_PATH
if it is not already defined and to export it.