Fix tricky lint check
Originally the code was $*, which triggered SC2048, that was changed to
$@ in 3040b79 (sharness.sh: fix SC2048: Use $@ (with quotes) to prevent
whitespace problems, 2018-09-04), but that fix is with "$*", not $@.

Changing it to $@ only triggers another issue SC2068, which is fixed
changing it to "$@".

None of those fixes are what we want, and in fact $@ behaves exactly the
same as $* (when unquoted).

This function is supposed to be called with arguments such as "FOO,BAR",
not "FOO BAR" so there's no need to bother thinking about $* or $@,
so change it to $1.

We don't want "$1" though, because we want the argument to be expanded
by IFS, so it's $1.

That creates the final problem of SC2086, so just silence that.

For reference:
https://lore.kernel.org/git/4A028A0A.5070003@viscovery.net/

Fixes SC2068.

Cc: Christian Couder <christian.couder@gmail.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>