Make lib-sharness optional
When the framework was split in 21fd2c0 (Move functions into
lib-sharness/functions.sh, 2019-04-26) essential functions were moved
into lib-sharness when that wasn't the intention of upstream.

The idea behind the move was to make something like this possible:

    test_perf 'foo' 'test_must_fail false'

For that test_must_fail should be in a different file.

However, that doesn't apply to essential functions like
test_expect_success which should always be available.

If essential functions are always available, then lib-sharness becomes
optional.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>