Fix zsh default srcdir support
It turns that by adding zsh support in 00cb93e (Add initial support for
zsh, 2023-03-10) we disabled the feature enabled in 0bc8705 (Set proper
location of TEST_SRCDIR, 2022-09-07) for zsh.

In order for SHARNESS_TEST_SRCDIR to have a proper default in zsh, we
need to restore the previous semantics of $0.

Let's do that by storing the value of $0 before POSIX_ARGZERO is turned
on.

Helps issue #124.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>